Polyfill `FindFirstFileExW()`
`std::filesystem::directory_iterator` calls this function with
`FindExInfoBasic`, which leaves some of the data in the returned
`WIN32_FIND_DATA` structure empty.
Note how we still assume that the MSLU hooks the W version of this
function. MSLU just doesn't replace `fInfoLevelId` because that flag is
newer.

Part of P0310, funded by Root and Arandui.